requirements.txt
Django
djangorestframework
django-statici18n
